+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Feat
 
+- **ic-asset-certification**: add chunkwise-handling of long assets with encodings
+- **ic-asset-certification**: add 206-chunking of long assets
 - **ic-http-certification**: add add_certificate_header util function
 - **ic-http-certification**: add common constructs to skip certification
 - **ic-http-certification**: add add_certificate_header util function
@@ -23,6 +25,11 @@
 
 ### Fix
 
+- clippy
+- clippy
+- address review feedback
+- clippy
+- clippy
 - clippy
 
 ### Refactor